add fees, benchmarking, valve and 20% limit

+++ pallets/automation-price/src/lib.rs
@@ -40,6 +40,9 @@ pub use pallet::*;
 mod benchmarking;
 pub mod weights;
 
+mod fees;
+pub use fees::*;
+
 use core::convert::TryInto;
 use cumulus_pallet_xcm::{Origin as CumulusOrigin};
 use frame_support::{
@@ -146,6 +149,9 @@ pub mod pallet {
 		/// The Currency type for interacting with balances
 		type Currency: Currency<Self::AccountId>;
 
+		/// Handler for fees
+		type FeeHandler: HandleFees<Self>;
+
 		type Origin: From<<Self as SystemConfig>::Origin>
 			+ Into<Result<CumulusOrigin, <Self as Config>::Origin>>;
 	}
@@ -222,6 +228,10 @@ pub mod pallet {
 		MaxTasksReached,
 		/// Failed to insert task
 		TaskInsertionFailure,
+		/// Insufficient Balance
+		InsufficientBalance,
+		/// Restrictions on Liquidity in Account
+		LiquidityRestrictions,
 	}
 
 	#[pallet::event]
@@ -278,7 +288,7 @@ pub mod pallet {
 		/// * `trigger_percentage`: what percentage task should be triggered at 
 		///
 		/// # Errors
-		#[pallet::weight(<T as Config>::WeightInfo::schedule_notify_task_full(1))]
+		#[pallet::weight(<T as Config>::WeightInfo::schedule_swap_task_extrinsic())]
 		#[transactional]
 		pub fn schedule_swap_task(
 			origin: OriginFor<T>,
@@ -288,7 +298,12 @@ pub mod pallet {
 			trigger_percentage: u128,
 		) -> DispatchResult {
 			let who = ensure_signed(origin)?;
-			Self::validate_and_schedule_task(who, provided_id, asset, direction, trigger_percentage)?;
+			let fee = <BalanceOf<T>>::saturated_from(1_000_000_000_000u64);
+			T::FeeHandler::can_pay_fee(&who, fee.clone())
+				.map_err(|_| Error::<T>::InsufficientBalance)?;
+			Self::validate_and_schedule_task(who.clone(), provided_id, asset, direction, trigger_percentage)?;
+			T::FeeHandler::withdraw_fee(&who, fee.clone())
+				.map_err(|_| Error::<T>::LiquidityRestrictions)?;
 			Ok(().into())
 		}
 
@@ -301,7 +316,7 @@ pub mod pallet {
 		/// * `directions`: number of directions of data input. (up, down, ?)
 		///
 		/// # Errors
-		#[pallet::weight(<T as Config>::WeightInfo::schedule_notify_task_full(1))]
+		#[pallet::weight(<T as Config>::WeightInfo::add_asset_extrinsic())]
 		#[transactional]
 		pub fn add_asset(
 			origin: OriginFor<T>,
@@ -312,6 +327,7 @@ pub mod pallet {
 			asset_owner: AccountOf<T>,
 			expiration_period: UnixTime,
 		) -> DispatchResult {
+			// TODO: needs fees if opened up to non-sudo
 			ensure_root(origin)?;
 			if expiration_period % 86400 != 0 {
 				Err(Error::<T>::InvalidAssetExpirationWindow)?
@@ -324,7 +340,7 @@ pub mod pallet {
 					upper_bound,
 					lower_bound,
 					expiration_period,
-					asset_sudo: asset_owner,
+					asset_sudo: asset_owner.clone(),
 				};
 				AssetMetadata::<T>::insert(asset.clone(), asset_metadatum);
 				let new_time_slot = Self::get_current_time_slot()?.saturating_add(expiration_period);
@@ -344,7 +360,7 @@ pub mod pallet {
 		/// * `directions`: number of directions of data input. (up, down, ?)
 		///
 		/// # Errors
-		#[pallet::weight(<T as Config>::WeightInfo::schedule_notify_task_full(1))]
+		#[pallet::weight(<T as Config>::WeightInfo::asset_update_extrinsic())]
 		#[transactional]
 		pub fn asset_update(
 			origin: OriginFor<T>,
@@ -358,12 +374,16 @@ pub mod pallet {
 					Err(Error::<T>::InvalidAssetSudo)?
 				}
 			}
+			let fee = <BalanceOf<T>>::saturated_from(1_000_000_000_000u64);
+			T::FeeHandler::can_pay_fee(&who.clone(), fee.clone())
+				.map_err(|_| Error::<T>::InsufficientBalance)?;
 			if let Some(asset_target_price) = Self::get_asset_target_price(asset.clone()) {
 				let last_asset_price: u128 = match Self::get_asset_price(asset.clone()) {
 					None => Err(Error::<T>::AssetNotSupported)?,
 					Some(asset_price) => asset_price,
 				};
-				let asset_update_percentage = Self::calculate_asset_percentage(value, asset_target_price).saturating_add(1);
+				let asset_update_percentage = Self::calculate_asset_percentage(value, asset_target_price)
+					.saturating_add(1);
 				let asset_last_percentage = 0;
 				if value > last_asset_price {
 					Self::move_scheduled_tasks(asset.clone(), asset_last_percentage, asset_update_percentage, 1)?;
@@ -371,6 +391,8 @@ pub mod pallet {
 					Self::move_scheduled_tasks(asset.clone(), asset_last_percentage, asset_update_percentage, 0)?;
 				}
 				AssetPrices::<T>::insert(asset.clone(), value);
+				T::FeeHandler::withdraw_fee(&who, fee.clone())
+					.map_err(|_| Error::<T>::LiquidityRestrictions)?;
 				Self::deposit_event(Event::AssetUpdated { asset });
 			} else {
 				Err(Error::<T>::AssetNotSupported)?
@@ -385,12 +407,13 @@ pub mod pallet {
 		/// * `directions`: number of directions of data input. (up, down, ?)
 		///
 		/// # Errors
-		#[pallet::weight(<T as Config>::WeightInfo::schedule_notify_task_full(1))]
+		#[pallet::weight(<T as Config>::WeightInfo::delete_asset_extrinsic())]
 		#[transactional]
 		pub fn delete_asset(
 			origin: OriginFor<T>,
 			asset: Vec<u8>,
 		) -> DispatchResult {
+			// TODO: needs fees if opened up to non-sudo
 			ensure_root(origin)?;
 			if let Some(_asset_target_price) = Self::get_asset_target_price(asset.clone()) {
 				AssetTargetPrices::<T>::remove(asset.clone());
@@ -411,10 +434,8 @@ pub mod pallet {
 		/// Complete as many tasks as possible given the maximum weight.
 		pub fn trigger_tasks(max_weight: Weight) -> Weight {
 			let mut weight_left: Weight = max_weight;
-			let run_task_weight = <T as Config>::WeightInfo::run_tasks_many_found(1)
-				.saturating_add(T::DbWeight::get().reads(1 as Weight))
-				.saturating_add(T::DbWeight::get().writes(1 as Weight));
-			if weight_left < run_task_weight {
+			let check_time_and_deletion_weight = T::DbWeight::get().reads(2 as Weight);
+			if weight_left < check_time_and_deletion_weight {
 				return weight_left
 			}
 
@@ -425,19 +446,25 @@ pub mod pallet {
 			};
 			if let Some(scheduled_deletion_assets) = Self::get_scheduled_asset_period_reset(current_time_slot) {
 				// delete assets' tasks
+				let asset_reset_weight = <T as Config>::WeightInfo::reset_asset(scheduled_deletion_assets.len().try_into().unwrap());
+				if weight_left < asset_reset_weight {
+					return weight_left
+				}
+				// TODO: this assumes that all assets that need to be reset in a period can all be done successfully in a block.
+				// 			 in the future, we need to make sure to be able to break out of for loop if out of weight and continue 
+				//       in the next block. Right now, we will not run out of weight - we will simply not execute anything if 
+				//       not all of the asset resets can be run at once. this may cause the asset reset triggers to not go off,
+				//       but at least it should not brick the chain.
 				for asset in scheduled_deletion_assets {
-					// delete asset tasks
 					Self::delete_asset_tasks(asset.clone());
-
-					// get time period duration
 					Self::update_asset_reset(asset.clone(), current_time_slot);
-					// 2. set new target price
 					if let Some(last_asset_price) = Self::get_asset_price(asset.clone()) {
 						AssetTargetPrices::<T>::insert(asset.clone(), last_asset_price);
 					};
 					Self::deposit_event(Event::AssetPeriodReset { asset });
 				}
 				ScheduledAssetDeletion::<T>::remove(current_time_slot);
+				weight_left = weight_left - asset_reset_weight;
 			}
 
 			// run as many scheduled tasks as we can
@@ -497,7 +524,7 @@ pub mod pallet {
 
 		pub fn run_notify_task(message: Vec<u8>) -> Weight {
 			Self::deposit_event(Event::Notify { message });
-			<T as Config>::WeightInfo::run_notify_task()
+			<T as Config>::WeightInfo::emit_event()
 		}
 
 		/// Runs as many tasks as the weight allows from the provided vec of task_ids.
@@ -513,7 +540,7 @@ pub mod pallet {
 				let action_weight = match Self::get_task(task_id) {
 					None => {
 						Self::deposit_event(Event::TaskNotFound { task_id: task_id.1.clone() });
-						<T as Config>::WeightInfo::run_tasks_many_missing(1)
+						<T as Config>::WeightInfo::emit_event()
 					},
 					Some(task) => {
 						let message = format!("asset: {:?}, percentage: {}, direction: {}", task.asset, task.trigger_percentage, task.direction);
@@ -527,7 +554,10 @@ pub mod pallet {
 
 				weight_left = weight_left.saturating_sub(action_weight);
 
-				if weight_left < <T as Config>::WeightInfo::run_tasks_many_found(1) {
+				let run_another_task_weight = <T as Config>::WeightInfo::emit_event()
+					.saturating_add(T::DbWeight::get().writes(1 as Weight))
+					.saturating_add(T::DbWeight::get().reads(1 as Weight));
+				if weight_left < run_another_task_weight {
 					break
 				}
 			}
@@ -614,10 +644,6 @@ pub mod pallet {
 			};
 			<Tasks<T>>::insert((asset, task_id), task);
 
-			// // This should never error if can_pay_fee passed.
-			// T::NativeTokenExchange::withdraw_fee(&who, fee.clone())
-			// 	.map_err(|_| Error::LiquidityRestrictions)?;
-
 			Self::deposit_event(Event::TaskScheduled { who, task_id });
 			Ok(())
 		}
@@ -629,7 +655,12 @@ pub mod pallet {
 			direction: u8,
 		) -> DispatchResult {
 			let mut existing_task_queue: Vec<(Vec<u8>, T::Hash)> = Self::get_task_queue();
-			for percentage in lower..higher {
+			// TODO: fix adjusted_higher to not peg to 20. Should move with the removal of 100 % increase.
+			let adjusted_higher = match higher > 20 {
+				true => 20,
+				false => higher,
+			};
+			for percentage in lower..adjusted_higher {
 				// TODO: pull all and cycle through in memory
 				if let Some(asset_tasks) = Self::get_scheduled_tasks((asset.clone(), direction.clone(), percentage.clone())) {
 					for task in asset_tasks {
@@ -652,6 +683,18 @@ pub mod pallet {
 			} else {
 				asset_update_value.saturating_sub(asset_target_price).saturating_mul(100).saturating_div(asset_target_price)
 			}
+		}
+	}
+
+	impl<T: Config> pallet_valve::Shutdown for Pallet<T> {
+		fn is_shutdown() -> bool {
+			Self::is_shutdown()
+		}
+		fn shutdown() {
+			Shutdown::<T>::put(true);
+		}
+		fn restart() {
+			Shutdown::<T>::put(false);
 		}
 	}
 }
